requirements.txt
anthropic
httpx
feedparser
beautifulsoup4
lxml
readability-lxml
jinja2
pydantic
python-dateutil
PyYAML
playwright
tenacity
pytest
python-dotenv
